+++ src/use-cases/scale.go
@@ -2,7 +2,7 @@ package usecases
 
 import "pdi/src/use-cases/process-image"
 
-func ScaleMatrix(matrix [][][]int, scale float64) [][][]int {
+func ScaleMatrix(matrix [][][]float64, scale float64) [][][]float64 {
 
 	transformationMatrix := &[3][3]float64{
 		{scale, 0, 0},
